Fix Giant Melee pitch (#14), Bunny Hood attachment (#15), and 64-bit safety bugs
- Fix #14: Change Player_GetMoreFlagsBit6 return type from bool to u8.
  Giant Melee sets more_flags.b6 = 2, which was previously truncated to 1 (bool),
  corrupting fp->x2228_b3 and forcing Tiny Melee pitch (+2 semitones) instead
  of Giant Melee low pitch.
- Fix #15: Load Bunny Hood ear offset vectors through DISC_VEC3_GET in ftCommon_8007FA00.
  Casting DiscVec3* to Vec3* previously stripped GCC big-endian scalar storage
  order, passing raw unswapped floats to it_80294EB0 (-430M units on Mario/Falcon,
  throwing ears into deep space, or ~0.0 on Fox/Sheik/Climbers, sinking ears into head).
- Fix 64-bit pointer truncation segfaults in HSD_SisLib_803A84BC, screenshot saves
  (hsd_80393A5C), Sheik chain joint creation (it_802BAF2C), and greens block logic.
- Clamp audio pitch ratio to 4.0f to avoid signed cvttss2si overflow UB on x86-64.
- Add DISC_STRUCT to snapshot header struct Unk80433380_0 for GC memory card saves.
- Fix Event Mode text color alpha channel reversal on little-endian platforms.
- Add bounds check in ftLib_80086630 for FTPART_INVALID.

+++ src/melee/lb/lbmemory.c
@@ -66,7 +66,7 @@ static inline Handle* new_handle(void* arenaLo, void* arenaHi)
     HSD_ASSERT(0x7B, _p(free_heap));
 
     if (PC_IS_ARAM_ADDR(arenaLo) && PC_IS_ARAM_ADDR(arenaHi)) {
-        HSD_ASSERT(0x80, (u32)arenaLo >= (u32)_p(a_arenaLo) && (u32)arenaHi <= (u32)_p(a_arenaHi));
+        HSD_ASSERT(0x80, (uintptr_t) arenaLo >= (uintptr_t) _p(a_arenaLo) && (uintptr_t) arenaHi <= (uintptr_t) _p(a_arenaHi));
     }
 
     POP_HANDLE(&_p(free_heap), h);
