+++ src/sysdolphin/baselib/hsd_3A76.c
@@ -450,7 +450,7 @@ static void sisFitLineToBox(HSD_Text* text, f32 measured_width)
     }
 }
 
-void HSD_SisLib_803A84BC(HSD_GObj* gobj, int pass)
+void HSD_SisLib_803A84BC(HSD_GObj* gobj, uintptr_t pass)
 {
     // clang-format off
     HSD_Text *text;
@@ -490,7 +490,7 @@ void HSD_SisLib_803A84BC(HSD_GObj* gobj, int pass)
         }
         text = HSD_GObjGetUserData(gobj);
     } else {
-        text = (HSD_Text*) (uintptr_t) (u32) pass;
+        text = (HSD_Text*) pass;
     }
     if (text->hidden == 0 && text->sis_buffer != NULL) {
         u8 *sis_cursor = (u8 *)text->sis_buffer;
